requirements.txt
discord.py
psycopg2
requests
asyncio